chore: Navigationのstyleを実装
