--- src/app/(components)/Navigation.tsx
@@ -1,22 +1,0 @@
-'use client';
-
-import Link from 'next/link';
-
-export function Navigation() {
-  return (
-    <ul>
-      <li>
-        <Link href='/'>Home</Link>
-      </li>
-      <li>
-        <Link href='/about'>About</Link>
-      </li>
-      <li>
-        <Link href='/products'>Products</Link>
-      </li>
-      <li>
-        <Link href='/contact'>Contact</Link>
-      </li>
-    </ul>
-  );
-}
